Build fix.
        * css/CSSParser.cpp:
        (WebCore::CSSParser::parseTimingFunctionValue):
        * css/CSSStyleSelector.cpp:
        (WebCore::CSSStyleSelector::mapTransitionTimingFunction):



Canonical link: https://commits.webkit.org/21150@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@27229 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ WebCore/css/CSSParser.cpp
@@ -2189,7 +2189,7 @@ bool CSSParser::parseTimingFunctionValue(ValueList*& args, float& result)
     Value* v = args->current();
     if (!validUnit(v, FNumber, strict))
         return false;
-    result = v->fValue;
+    result = (float)v->fValue;
     if (result < 0 || result > 1.0f)
         return false;
     v = args->next();
